#136 refactor: slack - 인터페이스 분리
